requirements.txt
cloudinary==1.22.0
Django==2.2.6
django-allauth==0.42.0
django-crispy-forms==1.9.2
oauthlib==3.1.0
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
sqlparse==0.3.1
stripe==2.50.0
